% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_define.R
\name{add_dataset}
\alias{add_dataset}
\title{Adds a dataset to define XML document}
\usage{
add_dataset(dfn, ds)
}
\arguments{
\item{dfn}{A definition object.}

\item{ds}{A dataset definition.}
}
\description{
Adds a dataset to define XML document
}
