% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_dataset.R
\name{add_variable}
\alias{add_variable}
\title{Add a variable to a dataset definition}
\usage{
add_variable(dst, name, label = NULL, width = NULL, type = NULL)
}
\arguments{
\item{dst}{The dataset definition to which this variable will be added.}

\item{name}{The name of the variable.}

\item{label}{The label for this variable.}

\item{width}{The width of this variable.}

\item{type}{The data type for this variable.}
}
\value{
The dataset spec with the variable appended.
}
\description{
Add a variable to a dataset definition
}
