% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adam_xml.R
\name{create_adam_xml}
\alias{create_adam_xml}
\title{Create XML for ADAM}
\usage{
create_adam_xml(lst, version)
}
\arguments{
\item{lst}{A list of data frames that contain SDTM metadata.}

\item{version}{The version of the define XML to create.  Currently
only 2.0.0 is supported, which is the default.}
}
\value{
A vector of XML strings.
}
\description{
Function generates XML for the ADAM define.xml file.
}
