% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sdtm_xml.R
\name{create_sdtm_xml}
\alias{create_sdtm_xml}
\title{Create XML for SDTM}
\usage{
create_sdtm_xml(lst, version = "2.0.0")
}
\arguments{
\item{lst}{A list of data frames that contain SDTM metadata.}

\item{version}{The version of the define XML to create.  Currently
only 2.0.0 is supported, which is the default.}
}
\value{
A list of tibbles.
}
\description{
Function generates XML for the SDTM define.xml file.
}
