% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-dictionary.R
\name{def_prep}
\alias{def_prep}
\title{Prepare dictionary structure from a data frame with columns identifying variable, value, and definition}
\usage{
def_prep(
  definitions_tbl,
  variable_col,
  value_col,
  definition_col,
  drop_solos = TRUE
)
}
\arguments{
\item{definitions_tbl}{A data frame containing information about the definitions of values stored in another table.}

\item{variable_col}{Unquoted name of column in `definitions_tbl` that identifies columns in the data table being cleaned}

\item{value_col}{Unquoted name of column in `definitions_tbl` that stores current values of variables identified by `variable_col`}

\item{definition_col}{Unquoted name of column in `definitions_tbl` that stores replacement / "definition" values matched to `value_col`}

\item{drop_solos}{Optionally, logical identifying whether to exclude variables from the dictionary if they only have one unique value}
}
\value{
Named list of vectors to use for recoding variables. Probably should be an S3 class.
}
\description{
value map will be applied only when starting_value and ending_value are same
to do this, convert it into a nested list
... list will have elements for columns in pums_hh, each one a named vector of potential values
}
\examples{
tibble::tibble(variable   = c("A", "A", "B", "B", "C"),
               value      = c(1,   2,   1,   2,   1),
               definition = c("cat", "dog", "fred", "george", "lonely")) \%>\%
  def_prep(variable, value, definition)
}
