% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-recode-all.R
\name{def_recode_all}
\alias{def_recode_all}
\title{Recode all columns in a dataset using a definitions list}
\usage{
def_recode_all(values_tbl, definitions)
}
\arguments{
\item{values_tbl}{A data frame with values stored in a coded format}

\item{definitions}{A named list of named vectors to use for recoding variables.}
}
\value{
A data frame with the same rows and columns as \code{values_tbl}, with columns recoded according to \code{definitions}
}
\description{
Recode all columns in a dataset using a definitions list
}
