% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definer-package.R
\docType{package}
\name{definer-package}
\alias{definer}
\alias{definer-package}
\title{definer: Makes it easy to apply field maps / survey data definitions to datasets with integer coding}
\description{
Survey data is often delivered with categorical variables coded as integers. 
    This saves a lot of space but can make analysis pretty painful. This package helps 
    you use a recode / definitions table to update values of variables painlessly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davisadamw/definer}
  \item Report bugs at \url{https://github.com/davisadamw/definer/issues}
}

}
\author{
\strong{Maintainer}: Adam Davis \email{davisadamw@gmail.com} (\href{https://orcid.org/https://orcid.org/0000-0001-6324-6380}{ORCID})

}
\keyword{internal}
