% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineR.R
\docType{package}
\name{defineR}
\alias{defineR}
\title{Creates Define XML documents}
\description{
The \strong{defineR} package contains functions to
create CDISC compliant define.xml documents suitable for
submission to the FDA. The package also contains related capabilities
to help create and manage metadata needed to create the define.xml.
}
\details{
The \strong{defineR} package helps create define.xml files for FDA
submission.  More here...
}
\section{How to use}{

There are three steps to creating a define.xml document:
\itemize{
  \item Import metadata
  \item Create XML
  \item Write out XML markup
}
In the \strong{defineR} package, the \code{\link{import_metadata}} function
reads in the metadata. Currently, the package supports spreadsheet-based
metadata. Once you have the metadata, you can create XML for an SDTM or ADAM
define.xml using \code{\link{create_sdtm_xml}} and
\code{\link{create_adam_xml}}.  Then write out the XML to a file using
\code{\link{write_markup}}.

The package also supports the ability to create an HTML page to explore
your define.xml.

See function documentation for additional details.
}

