% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definer-example.R
\name{definer_example}
\alias{definer_example}
\title{Get path to definer example data}
\usage{
definer_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
\code{definer} comes bundled with a number of example data definition files in its inst/extdata directory.
  This function make them easy to access. Adapted from \code{\link[readr]{readr_example}()}.
}
\examples{
definer_example()
definer_example("ca_vehicle_survey_2019_definitions.xlsx")
}
