% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metadata_xl.R
\name{import_metadata_xl}
\alias{import_metadata_xl}
\title{Imports metadata from an Excel file}
\usage{
import_metadata_xl(file_path)
}
\arguments{
\item{file_path}{The path to the Excel metadata file.}
}
\value{
A list of tibbles, one for each sheet in the excel file.
}
\description{
Imports define metadata from xls or xlsx files. Returns
a list of tibbles.
}
