% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_define.R
\name{print.dfn_spec}
\alias{print.dfn_spec}
\title{Print a define specification}
\usage{
\method{print}{dfn_spec}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{Object to print.}

\item{...}{Follow on parameters.}

\item{verbose}{Whether to print in verbose mode.}
}
\description{
Print a define specification
}
