% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_document.R
\name{print.doc_spec}
\alias{print.doc_spec}
\title{Print a define document specification}
\usage{
\method{print}{doc_spec}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{Object to print.}

\item{...}{Follow on parameters.}

\item{verbose}{Whether to print in verbose mode.}
}
\description{
Print a define document specification
}
