% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_define}
\alias{write_define}
\title{Writes a define XML document to the file system}
\usage{
write_define(
  path,
  dir = ".",
  type = "sdtm",
  ver = NULL,
  check = TRUE,
  html = TRUE,
  view = TRUE,
  report_type = "PDF"
)
}
\arguments{
\item{path}{The path to the metadata file.  Currently only Excel metadata
files are supported.}

\item{dir}{The output directory to create the define.xml and associated files.
By default, the current working directory will be used. If the directory
does not exist, the function will attempt to create it.}

\item{type}{The type of define.xml to create.  Valid values are
"sdtm" and "adam".  Default is "sdtm".}

\item{ver}{The version of the define XML to produce.  Default is the
preferred version of the FDA. Currently only version "2.0.0" is supported.}

\item{check}{Whether or not to perform consistency checks.  If so,
a check report will be produced in the output directory. Valid values
are TRUE, FALSE, and "V5".  Default is TRUE.  The "V5" option will include
checks that metadata conforms to SAS Transport file version 5 requirements.}

\item{html}{Whether or not to produce the HTML output associated with the
define.xml.  Valid values are TRUE and FALSE.  Default is TRUE.}

\item{view}{Whether or not to show the check report in the viewer.  Valid
values are TRUE and FALSE.  Default is TRUE.}

\item{report_type}{The output type of the check report, if requested.
Valid values are "TXT", "RTF", "PDF", "HTML" and "DOCX".  Default is
"DOCX".}
}
\value{
The define.xml file and any associated files will be written
to the directory specified on the \code{dir} parameter.
}
\description{
The \code{write_define} function inputs an SDTM or ADAM
metadata file and exports a define.xml and associated files to a specified
directory. Possible associated files are the HTML transformation of
the define.xml and a check report. By default, the check report will
also be shown in the viewer.
}
\details{
If there is
a serious error producing the XML, the XML and HTML outputs will not be
produced.  A check report will produced that shows the error.
}
\examples{
# Create data
}
