% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml.R
\name{write_xml}
\alias{write_xml}
\title{Writes a define XML document to the file system}
\usage{
write_xml(xml, file_path)
}
\arguments{
\item{xml}{The xml to write.}

\item{file_path}{The path to write the xml file.}
}
\description{
Writes a define XML document to the file system
}
