% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adam_xml.R
\name{create_adam_xml}
\alias{create_adam_xml}
\title{Imports metadata}
\usage{
create_adam_xml(lst, version)
}
\arguments{
\item{location}{The location of the metadata.  For a file data
source, this is the path to the file.}

\item{type}{The type of data source for the metadata.  Currently
only supports the "excel" source type, which is the default.}
}
\value{
A list of tibbles.
}
\description{
Imports define metadata from a data source. Returns
a list of tibbles.
}
