% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_define.R
\name{write_define}
\alias{write_define}
\title{Writes a define XML document to the file system}
\usage{
write_define(dfn)
}
\arguments{
\item{dfn}{The define object to write.}
}
\description{
Writes a define XML document to the file system
}
