% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_markup.R
\name{write_markup}
\alias{write_markup}
\title{Writes an markup document}
\usage{
write_markup(x, file_path)
}
\arguments{
\item{x}{The markup to write.}

\item{file_path}{The path to write the markup file.}
}
\value{
The full path to the file.
}
\description{
The function writes an XML or HTML document to the file system.
}
\seealso{
\code{\link{create_sdtm_xml}} and \code{\link{create_adam_xml}}
function to generate XML.
}
