% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{write_metadata}
\alias{write_metadata}
\title{Writes a define XML metadata spreadsheet to the file system}
\usage{
write_metadata(
  dir = ".",
  type = "sdtm",
  ver = NULL,
  src_dir = NULL,
  check = TRUE,
  demo = FALSE
)
}
\arguments{
\item{dir}{The directory in which to produce the metadata. The default
is the current working directory.}

\item{type}{The type of metadata file to produce.  Valid values are
"SDTM" and "ADAM". Default is "SDTM".}

\item{ver}{The version of the metadata file.  Default is
the preferred version of the define standard.  Function
currently supports only version 2.0.0.}

\item{src_dir}{If the metadata will be created from existing datasets,
supply the directory path for those datasets.  If the metadata will
not be created from datasets, this parameter should be NULL.}

\item{check}{Whether or not to run the checks.}

\item{demo}{If this parameter is TRUE, the function will generate
a metadata file with example data.  Default is FALSE.}
}
\value{
The path of the generated file.
}
\description{
The function generates metadata and
writes it to the file system. The function can generate a blank metadata
spreadsheet, a demonstration spreadsheet, or metadata generated
from existing data files.
}
\details{
Here are some details.
}
\examples{
# Example data
}
